Initial creation
This was done just as an exercise to annoy a specific hosting provider with a phonetically similar name which uses shills in an attempt to discount unhappy customers.

This simple script allows you to edit your user's crontab with a common 'crontab' command available with most versions of cron.  I do very simple (and redundant) checks just to make it not-bash compliant.